pyproject.toml
[tool.poetry]
name = "es-bgm"
version = "1.0.1"
description = "Allows you to add background music to EmulationStation"
authors = ["David <davigetto@gmail.com>"]
license = "LGPL-3.0-or-later"
packages = [
    { include = "bgm" }
]

[tool.poetry.dependencies]
python = "^3.8"
pygame = "^2.0.1"
python-decouple = "^3.4"
PyYAML = "^6.0"
deepmerge = "^1.0.1"
pydantic = "^1.9.0"
transitions = "^0.8.10"
confuse = "^1.7.0"
psutil = "^5.9.0"

[tool.poetry.dependencies.pylint]
version = "*"
optional = true

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
pytest-mock = "^3.6.1"
doublex = "^1.9.3"
bunch = "^1.0.1"
decorator = "^5.1.1"


[tool.poetry.extras]
lint = ["pylint"]

[tool.commitizen]
name = "cz_conventional_commits"
version = "1.0.1"
tag_format = "$version"
version_files = [
    "bgm/__version__.py",
    "pyproject.toml:version"
]
update_changelog_on_bump = true
annotated_tag = true

[tool.black]
line-length = 120

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
